+++ packages/core/src/Models/Product.php
@@ -116,7 +116,7 @@ class Product extends BaseModel implements SpatieHasMedia
         return $this->belongsToMany(
             Collection::class,
             config('getcandy.database.table_prefix').'collection_product'
-        )->withPivot(['position']);
+        )->withPivot(['position'])->withTimestamps();
     }
 
     /**