+++ packages/admin/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+- Added collections to the product editing screen.
+
 ## 2.0-beta9 - 2022-02-11
 
 ### Fixed